add JWT auth to delete book

+++ src/BackEnd/PageRhythm/routes/book_route.py
@@ -17,8 +17,8 @@ def create_book():
     if not data:
         return jsonify({"error": "Invalid input", "message": "No JSON data found"}), 400
     data['owner_id'] = owner_id
-    result = book_service.create_book(data)
     
+    result = book_service.create_book(data)
     if result:
         return jsonify({"message": "Book created successfully"}), 201
     return jsonify({"message": "Failed to create book"}), 400
@@ -27,6 +27,7 @@ def create_book():
 @book_blueprint.route('/<string:book_id>', methods=['GET'])
 def get_book_information(book_id):
     book = book_service.get_book_information(book_id)
+    
     if book:
         return jsonify(book.to_serializable_JSON()), 200
     return jsonify({"message": "Book not found"}), 404
@@ -45,17 +46,25 @@ def update_book(book_id):
     current_identity = json.loads(get_jwt_identity())
     owner_id = current_identity["account_id"] 
     data = request.json  
+    
     if not book_service.check_ownership(book_id, owner_id):
         return jsonify({"message": "You do not have permission to update this book"}), 403
-    result = book_service.update_book(book_id, data)
     
+    result = book_service.update_book(book_id, data)
     if result:
         return jsonify({"message": "Book updated successfully"}), 200
     return jsonify({"message": "Failed to update book"}), 400
 
 # 5. Delete a book
 @book_blueprint.route('/<string:book_id>', methods=['DELETE'])
+@jwt_required()
 def delete_book(book_id):
+    current_identity = json.loads(get_jwt_identity())
+    owner_id = current_identity["account_id"]  
+    
+    if not book_service.check_ownership(book_id, owner_id):
+        return jsonify({"message": "You do not have permission to delete this book"}), 403
+
     result = book_service.delete_book(book_id)
     if result:
         return jsonify({"message": "Book deleted successfully"}), 200
